Updated readme Added *./Renders/ to the gitignore Added ability to draw the grid Added ability to save the layouts to file Added mouse position on the map grid Added ability to resize the grid Greated a group box to display layout data, currently only shows occurance rate
Removed the old console application

+++ DigimonWorld2Tool/DigimonWorld2Tool/Utility/Vector2.cs
@@ -55,8 +55,8 @@ namespace DigimonWorld2MapVisualizer.Utility
 
         public static Vector2 operator +(Vector2 a, Vector2 b) => new Vector2(a.x + b.x, a.y + b.y);
         public static Vector2 operator -(Vector2 a, Vector2 b) => new Vector2(a.x - b.x, a.y - b.y);
-        public static Vector2 operator *(Vector2 a, byte b) => new Vector2(a.x * b, a.y * b);
-        public static Vector2 operator *(byte a, Vector2 b) => new Vector2(b.x * a, b.y * a);
+        public static Vector2 operator *(Vector2 a, int b) => new Vector2(a.x * b, a.y * b);
+        public static Vector2 operator *(int a, Vector2 b) => new Vector2(b.x * a, b.y * a);
 
         public static bool operator ==(Vector2 a, Vector2 b) => a.x == b.x && a.y == b.y;
         public static bool operator !=(Vector2 a, Vector2 b) => a.x != b.x || a.y != b.y;
